Refactor ProfileContributePage and update Footer component: Removed the ProfileContributePage component, updated import paths, and refactored the Footer component to use useEffect for QR code generation based on the showQrcode prop.
